configs/droid4: Add support for droid 4 kexecboot image
On droid 4, we have a 4MB partition on the eMMC that we can use
for a minimal buildroot rootfs for kexec. This partition is the
"bpsw" partition at mmcblk1p13 that seems to be completely unused.

Probably "bpsw" was at some point used for the baseband firmware,
but that no longer seems to be the case. I've verified this by
re-flashing the modem firmware and the buildroot image on
mmcblk1p13 did not get trashed.

In order to fit the buildroot rootfs to 4MB, we need to use
absolute path symlinks for busybox binaries. This is probably
because the droid 4 stock kernel uses selinux. And we cannot use
BR2_PACKAGE_BUSYBOX_INDIVIDUAL_BINARIES because it won't fit into
the 4MB size. So we need to fix up the busybox symlinks in the
post-build.sh script.

And to avoid any rootfs corruption, let's mount rootfs read-only.
For that to work with the droid 4 stock kernel, we need to make
/dev to use tmpfs with a custom fstab.

For the buildroot image, we can easily fit in WLAN support with
wpa_supplicant, ssh and autossh so the buildroot image can be
modified to enable networking for easy kernel updates for the
developers. Note that configuring WLAN for the stock kernel
will be added in a separate patch.

Signed-off-by: Tony Lindgren <tony@atomide.com>