package/droid4-modules: Add droid 4 kexec and uart modules
The stock kernel on droid 4 is signed and it's not possible to
upgrade it. In order to have uart and kexec support, we need
to add few kernel modules for it. And we need the uart
module loaded as early as possible for proper error messages.

As the droid 4 stock kernel modules need to be built with
gcc-4.8 against a hacked up v3.18 droid 4 kernel tree, let's
just download them. We don't want to limit the buildroot
rootfs to an old compiler.

Signed-off-by: Tony Lindgren <tony@atomide.com>