package/kexec-droid4: Add support for booting legacy kernels
The stock v3.0.8 based kernel on droid 4 uses a custom device tree format.
It's little-endian, and requires atags in addition to the dtb for kernel
command line and initramfs information. So let's just use the legacy patched
kexec as getting these features merged to the mainline kexec-tools is going
to be a pain and probably not worth the trouble. So for now, let's just
use a separate binary for the legacy kexec and use a wrapper shells script.

To save space, we might be able to rebase these changes on top of the
mainline kexec-boot, this would save about 100KB of disk space.

Note that currently the v3.0.8 based Android images have an issue where
they depend on SafeStrap configuration being on mmcblk1p25 for their
initramfs. If somebody has a chance, please update the Android images to
use proper kernel command line for the root partition if available. And
after that we should be able to boot v3.0.8 based Android images too.

This is based on the Hashcode moto-omap4 branch.