configs/droid4: Add preinit.sh for booting to stock Android and SafeStrap
We need to start kexecboot as PID 1 for pivot_root to stock Android and
SafeStrap to work. To do that, let's add /sbin/preinit.sh and start that
first.

This means we only continue booting to buildroot when user exits kexecboot,
or configures /etc/preinit to not start kexecboot at all for automated
scripts for downloading developer kernels over WLAN for example.

Then our kexec wrapper script can see if the pivot_root is requested, and
we can kill kexecboot and have preinit.sh do pivot_root to the stock Android
distro.

And also add a default /boot/boot.cfg with the stock Android boot entry with
lowest possible priority to have it always available.

And note that we currently leave the busybox to the stock kernel root into
/bin/busybox as it's way more usable than the one in the stock distro.

Also note that we set a default password based on the SoC die ID. Half of
the die ID is also used by Android as the USB peripheral serial number and
the rest seems less random so it's not very secure.

Finally please note that the utags partition needs to be updated to point
to the new /sbin/preinit.sh script.

Signed-off-by: Tony Lindgren <tony@atomide.com>