Add initial battery charging support with battd
Let's start stock Android battd using runbattd wrapper before we start
kexecboot. This allows battery to charge while kexecboot is running.
When stopping kexecboot, we stop battd, and if exited from kexecboot we
start battd again.

For now, to keep kexecboot running to charge, you can select System menu
-> Show debug info. There is currently no charge status support.

We can add a small app to show the charge status, or integrate support
into kexecboot. Note that space is a concern here as we need to fit
everything into the 4 MiB boot image.

If using a serial console to debug the bootloader, note that serial port
will not currently work after plugging in a USB charger.

Signed-off-by: Tony Lindgren <tony@atomide.com>