Fix pivot_root to stock android trashing the bootloader
If we fail to umount the bootloader rootfs after pivot_root to the
Android initramfs, bootloader won't work on the next boot as Android
seems to change the mounted bootloader executable permissions or
something like that.

Let's umount run and tmp so we can umount the old root after pivot_root.

Let's also change to use new_root instead of data name for the new root
to avoid confusion on what is mounted and why.

Fixes: f449ce30a9e4 ("Add initial battery charging support with battd")
Signed-off-by: Tony Lindgren <tony@atomide.com>